[sys_monitor] by Edison
更改错误的数据类型

+++ app/inc/args_parse.h
@@ -9,7 +9,7 @@ extern MonitorConfig *config;
 void init_default_config(MonitorConfig *config);
 // Parse command line arguments and update monitor configuration
 // Return 0 on success, -1 on invalid arguments
-s16 parse_args(s16 argc, s8 *argv[], MonitorConfig *config);
+s32 parse_args(s32 argc, s8 *argv[], MonitorConfig *config);
 void print_help();
 #ifdef __cplusplus
 }
