+++ app/inc/cpu_monitor.h
@@ -13,11 +13,11 @@ extern MonitorData *data;
 //   total: total CPU time (user+nice+system+idle+...)
 //   idle: CPU idle time
 // Return 0 on success, -1 on failure
-s16 read_cpu_raw_data(u64 *total, u64 *idle);
+s32 read_cpu_raw_data(u64 *total, u64 *idle);
 
 // Collect CPU usage data and update to MonitorData
 // Return 0 on success, -1 on failure
-s16 collect_cpu_data(MonitorData *data);
+s32 collect_cpu_data(MonitorData *data);
 #ifdef __cplusplus
 }
 #endif
