+++ app/inc/disk_monitor.h
@@ -20,7 +20,7 @@ extern "C"{
 //   disk_dev: target disk device name (e.g., sda)
 //   data: monitor data structure to store result
 // Return 0 on success, -1 on failure
-s16 collect_disk_data(const s8 *disk_dev, MonitorData *data);
+s32 collect_disk_data(const s8 *disk_dev, MonitorData *data);
 #ifdef __cplusplus
 }
 #endif
