+++ app/inc/load_monitor.h
@@ -9,7 +9,7 @@ extern MonitorData *data;
 // Collect system load average data from /proc/loadavg
 // Update 1/5/15 minute load to MonitorData
 // Return 0 on success, -1 on failure
-s16 collect_load_data(MonitorData *data);
+s32 collect_load_data(MonitorData *data);
 #ifdef __cplusplus
 }
 #endif
