+++ app/inc/mem_monitor.h
@@ -9,7 +9,7 @@ extern MonitorData *data;
 // Collect memory usage data from /proc/meminfo
 // Update data to MonitorData structure
 // Return 0 on success, -1 on failure
-s16 collect_mem_data(MonitorData *data);
+s32 collect_mem_data(MonitorData *data);
 #ifdef __cplusplus
 }
 #endif
